feature 2005.1- added some comments, renamed the battle footer file to replace the original battle foorter. Changed winner screen text to "YOU ARE THE WINNER!"
