+++ client/src/components/cards/BattleFooter.tsx
@@ -1,9 +1,10 @@
-import React from "react";
+
 import ActionButton from "../buttons/ActionButton";
 import { ActionState } from "/types/single/actionState";
 import socket from "../../socket";
 import { BaseCard } from "./BaseCard";
 import { GenericFooter } from "./GenericFooter";
+import React, { useRef, useState, useLayoutEffect } from 'react';
 
 interface BattleFooterProp{
     possibleActions: ActionState[];
@@ -11,75 +12,68 @@ interface BattleFooterProp{
 }
 
 export const BattleFooter = ({possibleActions, battleId}: BattleFooterProp) => {
+
     const handleActionClick = (action: ActionState) => {
         console.log(`Action selected: ${action}`);
-        // You can emit the selected action to the server here if needed
+
         socket.emit("action_selected", { action, battleId, playerId: socket.id });
     };
 
-    // const topButton=
-    // `
-    // justify-evenly
-    // inset-x-0
-    // flex 
-    // w-[92%]
-    // xl:w-[57.5%]
-    // items-center 
-    // bottom-[190px]
-    // xl:bottom-[170px]
-    // mx-auto
-    // fixed
-    // z-40
-    // `
-    const topButton=`
-        flex
-        items-center
-        justify-center
-        grid
-        grid-cols-2
-        gap-x-8
-        gap-y-15
-        xl:bottom-[90px]
-        pl-[1.4%]
-        fixed
-        z-40
+     const footer = 
     `
+    bg-[#FBD474]
+    mx-auto
+    rounded-tl-[5rem]
+    rounded-tr-[5rem]
+    border-[4px]
+    border-blackCurrant
+    border-b-0
+    flex 
+    inset-x-0
+    fixed
+    w-[95%]
+    xl:w-[60%]
+    bottom-0
+    z-[40]
+    `;
 
-    const footer = 
-        `
-        bg-[#FBD474]
-        mx-auto
-        font-[Jua]
-        w-115
-        h-30
-        rounded-tl-[1rem]
-        rounded-tr-[1rem]
-        border-[4px]
-        border-blackCurrant
-        border-b-0
-        flex 
-        flex-col
-        items-center
-        justify-center
-        text-wrap
-        pr-[1.5%]
-        pb-[10%]
-        pt-4
-        pb-10
-        inset-x-0
-        fixed
-        bottom-12
-        z-10
-        `;
+    const button=
+    `
+    flex 
+    place-items-center 
+    w-full 
+    justify-evenly 
+    grid 
+    grid-cols-2 
+    gap-y-[5rem] 
+    xl:gap-y-[4rem] 
+    pb-[2%]
+    
+    `
+    
+    // Get the height of the 4 buttons and make footer the 83% of that height
+    const gridRef = useRef<HTMLDivElement>(null);
+    const [footerH, setFooterH] = useState(0);
+    useLayoutEffect(() => {
+        if (!gridRef.current) return;
 
-    return(
-        <div className={`${footer}`}>
-            <div className={`${topButton}`}>
-                    {possibleActions.map((action, _) => (
-                        
-                        <ActionButton actionState={action} battleId={battleId!} />
+        const h = gridRef.current.getBoundingClientRect().height;
+
+        setFooterH(h * 0.83);
+    }, [possibleActions]);
+
+    return (
+        
+        <div className="fixed flex mx-auto flex-col inset-x-0 bottom-0 w-[95%] xl:w-[60%] justify-center">
+            <div ref={gridRef}>
+                <div className={button}>
+                    {possibleActions.map((action, i) => (
+                    <div className="z-[50]"><ActionButton key={i} actionState={action} battleId={battleId!} /></div>
                     ))}
+                </div>
+            </div>
+            <div style={{ height: `${footerH}px` }} className={footer}>
             </div>
         </div>
-    )
-}
+    );
+    };
